+++ semana15/projeto-sexta/src/index.ts
@@ -1,73 +1,38 @@
-import knex from 'knex';
+import {createTodoListUser, createTodoListTask, editTodoListUser, getTodoListUserById, getTodoListTaskById, app} from './functions';
+
 import express, { Request, Response } from "express";
-import dotenv from "dotenv";
-import { AddressInfo } from "net";
-import moment from 'moment'
 
-dotenv.config();
+//1. Criar usuário
 
-const connection = knex({
-  client: "mysql",
-  connection: {
-    host: process.env.DB_HOST,
-    port: Number(process.env.DB_PORT || "3306"),
-    user: process.env.DB_USER,
-    password: process.env.DB_PASSWORD,
-    database: process.env.DB_NAME,
-  },
+app.put("/user", async (req: Request, res: Response) => {
+    try {
+      await createTodoListUser(
+        req.body.name,
+        req.body.nickname,
+        req.body.email
+      );
+  
+      res.status(200).send();
+        console.log("Criado com sucesso!")
+    } catch (err) {
+      res.status(400).send({
+        message: err.message,
+      });
+    }
+  });
+
+  //2. Pegar usuário pelo id
+
+    app.get("/user/:id", async (req: Request, res: Response) => {
+    try {
+        const user = await getTodoListUserById(req.query.id as unknown as number);
+
+        res.status(200).send({
+        id: user,
+        });
+    } catch (err) {
+        res.status(400).send({
+        message: err.message,
+        });
+    }
 });
-
-const app = express();
-
-app.use(express.json());
-
-const server = app.listen(process.env.PORT || 3000, () => {
-if (server) {
-const address = server.address() as AddressInfo;
-console.log(`Server is running in http://localhost:${address.port}`);
-} else {
-console.error(`Failure upon starting server.`);
-}
-});
-
-const createTodoListUser = async (
-    name: string,
-    nickname: string,
-    email: string
-  ): Promise<void> => {
-    await connection
-      .insert({
-        name: name,
-        nickname: nickname,
-        email: email
-      })
-      .into("TodoListUser");
-  };
-
-const createTodoListTask = async (
-    title: string,
-    description: string,
-    limitDate: string,
-    creator_user_id: number
-  ): Promise<void> => {
-    await connection
-      .insert({
-        title: title,
-        description: description,
-        limit_date: moment(limitDate, "DD/MM/YYYY").format("YYYY/MM/DD"),
-        creator_user_id: creator_user_id
-      })
-      .into("TodoListTask");
-  };
-
-  try {
-    createTodoListTask(
-        "Comprar arroz",
-        "Ir no mercado",
-        "03/01/2021",
-        1
-        );
-  }
-  catch(err) {
-      console.log(err.message)
-  };
